back-end/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
pillow>=9.0.0
huggingface-hub>=0.16.0
pydantic>=2.0.0